refactor(#8): 패키지 명 변경
